Solve day 11b
The mathematics to use here was the fact that we know all the numbers that should be testable.
For multiplication operations, we actually only need to know the individual elemens of the prime factorization that are in the options. Doing modulo the multiplication of all of them is sure to preserve those.
When we sum, we also acutally only need to know that afterward the divisibiity of those given prime numbers doesn't change. But it does not.
To see this, we can think with the example of 1 prime number. Say number 5. If we want to know whether something is divisible, we want to know if it is in the prime factorization. Now multiply it a few times, it doesn't matter if we take the modulo 5 since fact whether 5 is in the factorization is maintained, in that case, we would always arrive at 0.
then add some values, now also you can see that we can better just work with the rest after dividing by 5, and that part will decide whether 5 is a prime factor.

Another way to see it is that he test always does the modulo anyway, and taking the module of a multiple of that can not hurt, so as long as we multiply all primes its ok to always take the modulo.